backend/requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
isort==5.12.0
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
